0.12-34 -> runtime configurable number of drives.
Changed the kernel module to default with one LUN (SMC).

The rc helper scripts parse the library_contents file to count the number
of configured drives and the create the required number of tape (SSC) devices.

There is a flag in the build_library_config 'USE_MIXED_DRIVES=0' will create
all drives as type IBM3580-TD3 otherwise the first three will be IBM, the next
three will be SDLT600 and the next as SONY-AIT4. Any further will default to
IBM 3580-TD3

Signed-off-by: Mark Harvey <markh794@gmail.com>