READ_ELEMENT_STATUS: Overhaul
In some places the voltag length was coded to 36, in others 32.
I added a VOLTAG_LEN symbol and used it everywhere for consistency.
In main() I added a simple if/then/else check
by library vendor and product, as Spectralogic libraries only return the
serial number in dvcid whereas other libraries return vendor, product, and
serial number.  We can extend this if/then/else to something more pretty in
the future if necessary if other library-specific checks are needed.

decode_element_status() could only handle one page descriptor, but in the case
of ANY there can be multiple pages in the result.  I reworked
decode_element_status() and dump_element_desc() to support multiple pages.

I reworked skel_element_descriptor so that it would work for all descriptor
types, and renamed it to fill_element_descriptor.  This got rid of a lot of
duplicate code for the different descriptor types.

Signed-off-by: Kevan Rehm <kfr@sgi.com>
Signed-off-by: Mark Harvey <markh794@gmail.com>