Understand multiple libraries on single host
etc/library_contents.sample
---------------------------
Remove the drive serial numbers in the sample library_config as
the field is no longer used, the serial number of each drive is
retrieved by the vtllibrary from the device.conf file.

etc/mhvtl.in
---------------------------
Change the default device.conf file created by mhvtl to use a
non-zero value for the library's message queue ID, as zero cannot
be used, it has a special meaning (read all queues).  Add the new
        Library ID: %ld Slot: %d
line for each sample drive in the default device.conf file.  Change
the script to verify the existance of the library_contents.%d file of
each configured library.  (Note that this is sloppy, the same
cartridges will get placed in each library if more than one library
exists in device.conf and the library_contents.%d files are missing.)
Fix stop and shutdown entry points to be able to stop multiple libraries.

usr/build_library_config
---------------------------
Change the script to use the device.conf file to configure the
mhvtl minor devices for both drives and libraries, as there can
be more than one library, and because the drive minor numbers
(message queue IDs) no longer match the drive's library slot location.

usr/dump_messageQ.c
---------------------------
Add a my_id global variable to contain the process's communication
queue ID.  It isn't really used, as dump_messageQ uses 0 as the
queue ID so that it can read all queues, but the loader expects the
my_id variable to exist.  Remove the incorrect Usage message, as
dump_messageQ doesn't take any parameters.

usr/make_vtl_devices
---------------------------
Change the script to use the device ID numbers for both libraries
and drives in the device.conf file to create the mhvtl minor
devices.  This is a one-to-one mapping, so knowing the minor device
number, you know the queue ID of the process that owns it.

usr/make_vtl_media.in
---------------------------
Change the script to iterate over all library_contents.%d files
that exist, as there may be more than one library.

usr/q.c and usr/q.h
---------------------------
Modify the format of inter-process messages to include the message
ID of the caller so that the receiver can then easily determine to
whom the reply should be sent.  Each process that will send messages
ID of the caller so that the receiver can then easily determine to
whom the reply should be sent.  Each process that will send messages
must define a global my_id value which is used as the sender ID value
in all sent messages.  Removed the word "priority" from various log
messages, as I found that confusing, because the message queue ID
really isn't being used as a priority, it's a simple ID.

usr/vtlcmd.c
---------------------------
Add a my_id global variable to contain the process's communication
queue ID.  vtlcmd will use a hard-coded VTLCMD_Q value for now, which
means only one vtlcmd can be run at a time.  Added some type definitions
so that vtlcmd can track whether the current command is for a library
or for a drive.  Changed the command line syntax so that libraries must
be provided a queue ID just like the drives.  When given a queue ID,
vtlcmd will now read the device.conf file to verify that a device with
that queue ID exists, and it will use the type of that device to verify
that the specified command is appropriate for that device type.

usr/vtllibrary.c
---------------------------
Add a my_id global variable to contain the process's communication
queue ID.  Added a drv_id to the d_info structure containing the
queue ID for that drive, which can now be different than the drive's
slot location.  Changed send_msg() calls to use the drive's drv_id.
Added a -q option similar to the vtltape version so that each vtllibrary
can be given it's own message queue ID at startup.  Changed
check_tape_load() to use the global my_id instead of the hard-coded
LIBRARY_Q value.  Passed the q_msg struct into routines list_map,
load_map, open_map, close_map, and empty_map so that they could look
up the sender's message queue ID and use it in the reply.  Modify
processMessageQ() to take the new q_msg struct as its first input
parameter, so that it has access to the id of the sender and can
use it in its reply.  Modify update_drive_details() to parse the
new drive device.conf line:
        Library ID: %ld Slot: %d
to get each drive's message queue ID.  Rework init_slot_info() so
that the vtllibrary looks up its library_contents.%d file using
its own message queue ID, so that each library can have its own
library_contents file.  Reworded some error messages to remove the
word "priority", as the queue ID is not really being used as a
priority, it is being used as a simple ID value.  I found the word
"priority" to be confusing.  Got rid of the minor variable, as the
my_id value is the minor value.

usr/vtltape.c
---------------------------
Add a my_id global variable to contain the process's communication
queue ID.  Modify processMessageQ() to take the new q_msg struct as
its first input parameter, so that it has access to the id of the sender
and use it in its reply.  Change main() to set my_id based upon the
-q command line option.  Reworded some error messages to remove the
word "priority", as the queue ID is not really being used as a
priority, it is being used as a simple ID value.  I found the word
"priority" to be confusing.  I changed all message senders to ensure
that they end each text message with a NULL character, allowing me
to then remove the forcing of a NULL byte in the receiver processing.

Signed-off-by: Kevan Rehm <kfr@sgi.com>
Signed-off-by: Mark Harvey <markh794@gmail.com>