Improve tape load/unload handling
Instead of relying on a single bit flag if tape is 'in mouth of drive',
save the barcode of the media. This way, any 'recovery' will
be easier to re-mount the media (we'll know what files to open).
Set 'barcode' to null if tape is moved out of drive.

e.g.
    mtx -f /dev/sg* load 1 0 (load tape into drive)
    mt -f /dev/nst* offline (eject the drive)
    mt -f /dev/nst* load (Now try to reload the tape)

Todo: Add an option to vtlcmd to remove the tape from the mouth
of a stand-alone drive.

Signed-off-by: Mark Harvey <markh794@gmail.com>