DIV: Write block of data on LBP mis-compare.
Mainly for debugging purposes. At least with the block of data - please
share the virtual media contents and the /var/log/messages for further
debugging.

The syslog will contain the RS-CRC/CRC32C value, and the contents of the
block of data will be written. Between the two, sould be able to
figure out what is being mis-calculated.

Signed-off-by: Mark Harvey <markh794@gmail.com>