systemd device config: touch up error logging
A follow-up to f4fbef, when mhvtl-device-conf-generator is run
by systemd (as it normally would be), messages to stderr are
not logged anywhere (as far as I can tell), but klog is preserved.

Introduce a very minimal shim writing messages to klog, with minimal
error checking and recovery on a best-effort basis (as there really
isn't much to do on failures at this stage). Make error opening
the config file log the fact to klog using this shim (the net effect
being there now is an easy-to-find trace of what'd happened
in dmesg and other places).