kernel: Re-work compat symbols detection
In the day and age of distros backporting various things from later
kernels to their shipped kernels, relying on kernel version numbers
for knowing whether a certain feature is available is not really
relyable.

Introduce a new auto-configure script that tries a little harder
to figure out whether some features (implemented in backports.h)
is necessary for a particular scenario. The method used is not
infallible and there probably is a more correct way, but it is
simple enough to solve the immediate problem.

Addresses and should fix #95