Fix kernel issue with queuecommand_lck
Newer kernel removed the second argument from
the SCSI mid-layer queue command function (the
"done" argument was removed). We were handling
that via checking the kernel version, but some
distros have back-ported the change to earlier
kernels, so just check to see if the function
prototype has one argument or two.

Also, clean up the queue command function, so
it's a little easier to read.