vtltape & vtllibrary: Change exit strategy
Wait for the 'backoff' to increment - indicating no I/O for a little
while. Then exit.

This works around the current behaviour where the slave_destroy() can
be called even if there is any I/O in flight - generally resulting
in a kernel panic in the slave_destroy() scsi mid-layer.

Side affect is the requirement for a lock file. Without a lock, there
is the potential for a new daemon to start while the previous one is
still active. Startup now waits up to 10 seconds for the lock file to
become available.

Reported-by: https://github.com/markh794/mhvtl/issues/131
Signed-off-by: Mark Harvey <markh794@gmail.com>