mhvtl: kernel: fix build issue if RealTime kernel
If CONFIG_PREEMPT_RT is defined, as it is for the
new SUSE kernel, usage of the __SPIN_LOCK_UNLOCKED() macros
fails.

It looks like this macro was always used incorrectly, but
it only gets caught if Real Time is enabled in the kernel.

The __SPIN_LOCK_UNLOCKED() macros is meant to be used for
variable static initialization, not for dynamic use, as
can be seen elsewhere in the kernel.

Fix this by using spin_lock_init(), which is meant for
dynamic use.